+++ packages/brain/src/calls/index.ts
@@ -2,13 +2,10 @@ export {
   beaconchaFetchAllValidatorsInfo,
   beaconchaFetchValidatorsInfo,
 } from "./beaconchaApi.js";
-export {
-  deleteValidators,
-  getValidators,
-  importValidators,
-  updateValidators,
-  exitValidators,
-  getExitValidators,
-} from "./validators.js";
+export { deleteValidators } from "./deleteValidators.js";
+export { exitValidators, getExitValidators } from "./exitValidators.js";
+export { importValidators } from "./importValidators.js";
+export { updateValidators } from "./updateValidators.js";
+export { getValidators } from "./getValidators.js";
 export { signerGetStatus } from "./signerGetStatus.js";
 export { getStakerConfig } from "./getStakerConfig.js";
